+++ backend/canisters/local_user_index/api/src/lib.rs
@@ -4,7 +4,7 @@ use types::nns::CryptoAmount;
 use types::{
     CanisterId, ChannelLatestMessageIndex, ChatId, ChitEarnedReason, CommunityId, Cryptocurrency,
     DiamondMembershipPlanDuration, MessageContent, MessageContentInitial, MessageId, MessageIndex, PhoneNumber, ReferralType,
-    SuspensionDuration, TimestampMillis, UniquePersonProof, UpdateUserPrincipalArgs, User, UserId,
+    SuspensionDuration, TimestampMillis, UniquePersonProof, UpdateUserPrincipalArgs, User, UserId, UserType,
 };
 
 mod lifecycle;
@@ -72,7 +72,10 @@ pub struct UserRegistered {
     pub user_id: UserId,
     pub user_principal: Principal,
     pub username: String,
+    #[serde(default)]
     pub is_bot: bool,
+    #[serde(default)]
+    pub user_type: UserType,
     pub referred_by: Option<UserId>,
 }
 
@@ -161,10 +164,14 @@ pub struct DeleteUser {
 pub struct GlobalUser {
     pub user_id: UserId,
     pub principal: Principal,
+    #[serde(default)]
     pub is_bot: bool,
+    #[serde(default)]
     pub is_platform_moderator: bool,
     pub diamond_membership_expires_at: Option<TimestampMillis>,
     pub unique_person_proof: Option<UniquePersonProof>,
+    #[serde(default)]
+    pub user_type: UserType,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]