+++ backend/canisters/local_user_index/impl/src/updates/c2c_notify_user_index_events.rs
@@ -73,7 +73,7 @@ fn handle_event(event: Event, state: &mut RuntimeState) {
             );
         }
         Event::UserRegistered(ev) => {
-            state.data.global_users.add(ev.user_principal, ev.user_id, ev.is_bot);
+            state.data.global_users.add(ev.user_principal, ev.user_id, ev.user_type);
 
             if let Some(referred_by) = ev.referred_by {
                 if state.data.local_users.get(&referred_by).is_some() {