+++ backend/canisters/user_index/impl/src/updates/c2c_notify_events.rs
@@ -10,7 +10,7 @@ use local_user_index_canister::{
     UserJoinedGroup, UserRegistered, UsernameChanged,
 };
 use storage_index_canister::add_or_update_users::UserConfig;
-use types::{CanisterId, MessageContent, TextContent, UserId};
+use types::{CanisterId, MessageContent, TextContent, UserId, UserType};
 use user_index_canister::c2c_notify_events::{Response::*, *};
 use user_index_canister::Event;
 
@@ -130,7 +130,7 @@ fn process_new_user(
     state
         .data
         .users
-        .register(caller, user_id, username.clone(), now, referred_by, false);
+        .register(caller, user_id, username.clone(), now, referred_by, UserType::User);
 
     state.data.local_index_map.add_user(local_user_index_canister_id, user_id);
 
@@ -140,6 +140,7 @@ fn process_new_user(
             user_principal: caller,
             username: username.clone(),
             is_bot: false,
+            user_type: UserType::User,
             referred_by,
         }),
         Some(local_user_index_canister_id),